+++ src/main/java/chess/domain/Color.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 public enum Color {
     BLACK,
