+++ src/test/java/chess/domain/PositionTest.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
