+++ src/test/java/techcourse/fp/mission/CarTest.java
@@ -1,8 +1,8 @@
 package techcourse.fp.mission;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 class CarTest {
