+++ src/test/java/techcourse/fp/mission/StreamStudyTest.java
@@ -1,10 +1,11 @@
 package techcourse.fp.mission;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.List;
-import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 class StreamStudyTest {
